+++ financial-goal-service/src/main/java/com/financial_goal_service/app/entity/FinancialGoal.java
@@ -14,8 +14,8 @@ public class FinancialGoal {
     private Long goalId;
     private Long custId;
     private String goalName;
-    private Integer targetAmount;
-    private Integer currentAmount;
+    private Double targetAmount;
+    private Double currentAmount;
     private Date targetDate;
     private String riskTolerance;
     private String status;
